+++ js/util.js
@@ -1,0 +1,20 @@
+const getRandomInteger = (min, max) => Math.floor(Math.random() * max - min + 1) + min;
+
+const createRandomIdGenerator = (min, max) => {
+  const previousValues = [];
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+export {getRandomInteger, createRandomIdGenerator};
+
